refactor WorkerSet into multisig
