+++ contracts/multisig-prover/src/contract.rs
@@ -101,12 +101,11 @@ mod tests {
         Addr, Fraction, Uint256, Uint64,
     };
     use cw_multi_test::{AppResponse, Executor};
-    use multisig::msg::Signer;
+    use multisig::{msg::Signer, workerset::WorkerSet};
 
     use crate::{
         encoding::Encoder,
         msg::{GetProofResponse, ProofStatus},
-        state::WorkerSet,
         test::{
             mocks,
             multicontract::{setup_test_case, TestCaseConfig},
