+++ contracts/multisig/Cargo.toml
@@ -50,6 +50,7 @@ rewards = { workspace = true, features = ["library"] }
 schemars = "0.8.10"
 serde = { version = "1.0.145", default-features = false, features = ["derive"] }
 serde_json = "1.0.89"
+sha3 = { version = "0.10", default-features = false, features = [] }
 thiserror = { workspace = true }
 
 [dev-dependencies]
